cleanup features
- move Protocol in basetypes
- replace Connection build_command method with Command build method
