+++ obdii/__init__.py
@@ -1,8 +1,8 @@
-__title__ = 'obdii'
-__author__ = 'PaulMarisOUMary'
-__license__ = 'MIT'
-__copyright__ = 'Copyright 2025-present PaulMarisOUMary'
-__version__ = '0.0.3a0'
+__title__ = "obdii"
+__author__ = "PaulMarisOUMary"
+__license__ = "MIT"
+__copyright__ = "Copyright 2025-present PaulMarisOUMary"
+__version__ = "0.0.3a0"
 
 from .connection import Connection
 from .commands import commands
